+++ CorgPhish/popup/main.js
@@ -469,6 +469,8 @@ const handleSettingsChange = async () => {
     warnOnUntrusted: dom.alertInput?.checked ?? DEFAULT_SETTINGS.warnOnUntrusted,
     linkHighlightEnabled:
       dom.linkHighlightToggle?.checked ?? DEFAULT_SETTINGS.linkHighlightEnabled,
+    antiScamBannerEnabled:
+      dom.antiScamToggle?.checked ?? DEFAULT_SETTINGS.antiScamBannerEnabled,
     strictMode: dom.strictModeToggle?.checked ?? DEFAULT_SETTINGS.strictMode,
     theme: dom.themeToggle?.checked ? "light" : "dark",
     language: dom.languageSelect?.value ?? DEFAULT_SETTINGS.language,
@@ -495,6 +497,9 @@ const updateSettingsControls = () => {
   }
   if (dom.linkHighlightToggle) {
     dom.linkHighlightToggle.checked = currentSettings.linkHighlightEnabled;
+  }
+  if (dom.antiScamToggle) {
+    dom.antiScamToggle.checked = currentSettings.antiScamBannerEnabled;
   }
   if (dom.strictModeToggle) {
     dom.strictModeToggle.checked = currentSettings.strictMode;
@@ -686,6 +691,7 @@ safeAddEvent(dom.clearHistoryBtn, "click", async () => {
 safeAddEvent(dom.autoCheckInput, "change", handleSettingsChange);
 safeAddEvent(dom.alertInput, "change", handleSettingsChange);
 safeAddEvent(dom.linkHighlightToggle, "change", handleSettingsChange);
+safeAddEvent(dom.antiScamToggle, "change", handleSettingsChange);
 safeAddEvent(dom.strictModeToggle, "change", handleSettingsChange);
 safeAddEvent(dom.themeToggle, "change", handleSettingsChange);
 safeAddEvent(dom.languageSelect, "change", handleSettingsChange);